Fix storage of Todoist's key event handlers #112
I noticed a few days ago that Todoist was setting document level
keyboard event handlers after page load.  My solution to that was to
frequently overwrite these handlers with the ones needed by
todoist-shortcuts.

I left the logic for storing Todoist's key event handlers the same
though, doing it on initialization.  It turns out this caused the
fallthrough to Todoist's handlers to fail.

The solution is to write down the old handler every time it gets
overwritten.